From Issue #50
We identified a bug in this module where the cache_lifetime of the Fastly_CDN_Block_Cookie_Formkey block is set to false instead of null which causes the block to be cached indefinitely by Magento's block caching mechanism instead of never being cached with the null setting.

This causes a conflict on the very first (and only the first) page load of any session where Magento will grab the cached version of the block, which contains a formkey from another session. This causes the ESI replacement logic to fail to replace this formkey value with an tag when the final page is sent to the user by Fastly.

The below patch remedies this by changing the block to never cache, which allows the \Fastly_CDN_Model_Observer::replaceFormKeys logic to work correctly 100% of the time, even on the first page load of a session by insuring that the contents of the block will always contain the current session's formkey value.